Ticket #1247: minor fix

+++ samples/ClientServer/demo_push_server/push_server_lib/internal/sopc_trustlist_helper.c
@@ -491,7 +491,7 @@ static SOPC_ReturnStatus gen_handle_with_retries(SOPC_TrustListContext* pTrustLi
     }
     else
     {
-        *deep = *deep - 1;
+        *deep = (uint8_t)(*deep - 1);
     }
     SOPC_TrLst_Handle genHandle = SOPC_TRUSTLIST_INVALID_HANDLE;
     SOPC_ExposedBuffer* pBuff = NULL;
